C4SequenceNumber is now an enum class (when using C++ API)
When LITECORE_CPP_API is defined as 1 — which is true when you include
c4Base.hh, and I’m now predefining it in the LiteCore build — the
type C4SequenceNumber is now a C++ `enum class` instead of a simple
typedef of `uint64_t`. So is the internal `sequence_t`, which is now
simply a typedef of `C4SequenceNumber`.

This means it’s no longer interchangeable with an integer at all.
You have to cast it to convert to or from an integer.

There are operator overloads that allow
* Adding/subtracting an integer to a C4SequenceNumber
* Subtracting two C4SequenceNumbers, producing a uint64_t
* Incrementing or decrementing a C4SequenceNumber

There’s also a C++ literal suffix `_seq` that when appended to an
integer literal makes it a C4SequenceNumber.