CBL-219: Move deleted docs to a different SQLite table (#1343)
* Deleted docs live in a separate SQLite table now

Optimization that removes deleted docs from the `kv_default` table
and stores them in a `kv_dead` table instead. This improves performance
in the default table since there are fewer rows, and since queries
don’t have to check whether a doc is deleted.

* Fixed querying deleted docs

Now that deleted docs are in a separate table, they can't be queried
the way they were before. Instead, QueryParser now checks whether the
WHERE clause matches _only_ deleted docs; if so, it operates on the
kv_deleted table instead, and the `_deleted` property becomes a no-op
(evaluates to `true`.) Otherwise, any use of the `_deleted` property
is an error.
This is a bit more limited since you can't write a query that will
match both deleted and non-deleted docs, but it should handle the
typical use cases.

* BothKeyStore improvements, unit tests

* Optimized recordCount(): it doesn't have to filter out tombstones
* Rewrote set() to be clearer
* Enumerator works in descending and unordered modes
* Enumerator now handles the case where a doc can be in both stores
* Improved unit tests

* Better db schema upgrade code, and tests

* Fixed regressions, post-merge, except for querying deletions

* Queries on a deleted-docs table

As long as they query for _only_ deleted docs.

* All unit tests pass w/deleted keystores

* Some cleanup of the last commit

* regenerate docs with doxygen

Co-authored-by: Jens Alfke <jens@couchbase.com>