CBL-2611: Take steps to get a reasonable backtrace on Linux (#1365)
The export symbol hiding in LiteCore prevents getting most function names, but instead an offset that can be used with addr2line can be printed.  However, this relies on correcting the stripping script in order to extract all needed symbols before the static libraries are stripped.  This is also corrected in this commit.  Additionally, in order to get usable function names out of the tests add -export_dynamic to their linker arguments.