Restored some Puller logs that got lost in a merge
Portions of commit 32c958de got lost in a merge I did right afterward
into the feature/pusher-refactor branch. The original commit was on
master, I merged it into the branch, and then later the branch got
merged back into master. This happened in July 2020.

Most of that commit was fixes to code that I'd reimplemented
on the branch, but there were some useful logging bits that I seem to
have missed and didn't propagate.

This commit restores those logs, except for one bit: Puller can't
log `_waitingChangesMessages.size` because it no longer exists; that
object is now in RevFinder and Puller can't access it.