CBL-2834 : Merge hotfix 2.8.7 (#1384)
* Add a test for the new fail-fast mode for database delete

If the caller (i.e. platform / application) is keeping the database open, rather than internal logistics, then new behavior is going to be implemented to fail immediately instead of waiting (similar to transient vs permanent network errors).  Implementation in a future commit.

* Tagging the opened database connections. CBL-2357 (#1253)

Tagging the opened database connections. CBL-2357
By default, a database connection is tagged "appOpened" as it is opened. These database connections are opened by the application, and, as such, the application must close all of them as it calls to delete the underlying database. Other tags are reserved for the connections opened by the Core library. The library is responsible to close them.
Also changed kOtherDBCloseTimeoutSecs from 3 seconds to 6 seconds for CBL-2383.

* Update for new sdkmanager location (#1231)

* Update for new sdkmanager location

This is the result of the docker container update to JDK 11.  The sdkmanager in the previous location is not compatible with JDK 11, and frustratingly the new sdkmanager will download the old sdkmanager into the old location so we are forced to use the new one.

* Update for PR validation as well

* Run Android validation on docker container

* CBL-2407: Tear down DBAccess when replicator stopped

As opposed to waiting until the replicator is completely out of memory.  This gives a more deterministic end of life to the object.

* Finer database tags to help debugging. CBL-2357 (#1256)

* Fix instability caused by fix for 2407

Now with the DBAccess closing earlier, straggling actor messages could access them.  Switch this to throw an error instead that won't crash the process.

* 2.8.7 vf pr, this PR (cherry) picks the resolutions of CBL-2410, CBL-2304 (#1257)

* CBL-2410: Don't tear down response timer in timeout situation

If onClose is reached with _timedOut being true, then that means not only did the timer already fire and tearing it down is extraneous at this point, but it is almost certainly the case that this method is reached synchronously from the timer callback via timer -> close/requestClose -> platform -> onClose.  This means that it is a hang situation in which the timer waits forever for its callback to finish, and the callback is waiting for the current method to finish.

* We need this fleece commit for CBL-2304

Co-authored-by: Jim Borden <jim.borden@couchbase.com>

Co-authored-by: Jim Borden <jim.borden@couchbase.com>