Result::value() now throws the error as an exception
It used to just throw AssertionFailed, but it makes much more sense
to throw the error it holds, and that might sometimes even be useful
if you use it in a function that throws exceptions.

Also, `_value` is a faster version of `value` that doesn't check for
errors, so it'll instead crash with a null-deref. Use it when you
already checked there's no error, and need the speed.