CBL-2764 : fix a bug of "heap-use-after-free". (#1398)
* CBL-2764 : fix a bug of "heap-use-after-free".
From the report of Mac's Address Sanitizer, we ran into a situation whereby, concurrently,
A. as the holder of the last reference of a Query object, LiveQuerierDelegate, is deleted, the Query object goes to the deletion process;
B. as DataFile is deleted, the close method is called on the same Query object.
For good reason, DataFile keeps plain pointers of Query objects, which are registered/unregistered as they are constructed/destructed. We fix the problem by protecting _queries in DataFile with mutex. We also need to make sure that the first thing in the destruction process of the Query object is to unregister it.