Change HTTP response for compatibility w/ Safari (#1403)
Safari refuses to make a WebSocket client connection (i.e. with a JS WebSocket
object) to LiteCore -- it logs "the server sent an illegal response" and aborts
the connection. No other client I've tried has such an issue.

I finally figured out that what triggers this is that the response starts
"HTTP/1.0". Changing it to "HTTP/1.1" fixes the problem.