CBL-2670 : ensure message order for LoopbackWebSocket. (#1409)
CBL-2670 : ensure message order for LoopbackWebSocket.
LoopbackWebSocket implements messsaging within the same application by threaded queues. For Mac, in particular, we use GCD's dispatch_after in order to simulate network latency. Since dispatch_after takes absolute time in the future, as opposed to a delta delay, the order may be displaced if two future times are numerically equal. To guarantee the ordering, we take advantage of the ordering guarantee of dispatch_async with the serial queue: we queue the messages with dispatch_async, and actually send the message with dispatch_after. We do two dispatches only if the latency is greater than 0.