Fix SequenceTracker bug with docs observed before a db change
SequenceTracker::_documentChanged was incorrectly ignoring a change
to a document with an existing-but-idle Entry when there were no db
change notifiers.

In practice this meant that if you created a doc-change notifier for
a doc that hadn't been modified yet in this db instance, and then
updated the document, the change wouldn't be recorded in the
SequenceTracker so other db instances wouldn't find out about it.

The offending lines were last edited in 2016(!), so I am not sure
exactly why they were there. I think they date back to a time when
there was only a single SequenceTracker shared by all DB instances on
a file, in which case that behavior would be a valid optimization.

It should be safe to backport this fix to any earlier version of
CBL, back to 2.0.

CBL-3036

update the Jenkinsfile.