Made C4ConnectedClient with with BuiltInWebSocket
The way the WebSocket was created wasn't compatible with the built-in
WebSocket implementation used by CBL-C and the cblite tool, so I
changed the call from `new C4SocketImpl` to `CreateWebSocket`.

However, this function required a C4Database pointer ... turns out
that's only so the WebSocket can access cookies. I changed the code
to allow a null pointer, in which case it just won't do any cookie
stuff. This is a short-term fix; long-term we need some other way to
persist cookies, I think.