CBL-2773 : use WeakHolder for websocket::Delegate. (#1471)
CBL-2773 : use WeakHolder for websocket::Delegate.
Following is a chain of strong ownership:
  Replicator -> Connection -> BLIPIO -> WebSocket
Currently, the reverse relationship from WebSocket to BLIPIO is established by plain pointer, for good reasons. However, this creates a problem if the websocket outlives BLIPIO for a fleeting moment of time. In this commit, we use WeakHolder for the delegate.