CBL-3224 : use WeakHolder for websocket::Delegate. (#1471) (#1481)
Following is a chain of strong ownership:
  Replicator -> Connection -> BLIPIO -> WebSocket
Currently, the reverse relationship from WebSocket to BLIPIO is established by plain pointer, for good reasons. However, this creates a problem if the websocket outlives BLIPIO for a fleeting moment of time. In this commit, we use WeakHolder for the delegate.
Cherry-picked from 7d8a12b7713aed