Remove extra isValid flag and reverse order of invalidation
Nulling the database on an invalid collection will do no harm since it doesn't own the database memory and won't affect its validity by setting it to null the way some of the other pointers will.  Since the collection can no longer be resurrected, it will never need to use the database again after that.  Furthermore, now that the definition of valid has changed, stop allowing operations after invalidate is called.  Previously keyStore and documentFactory were continued to be allowed as long as they were non null, but check isValid instead.  This avoids getting a cryptic SQLite misuse error instead of a LiteCore not open error.

In order to avoid weird situations where a creation run very shortly after deletion will not be successful, reverse the order of cache invalidation.  When a collection is deleted, inform the external files first and let them invalidate their entries and reset their data files to nonexistent.  Then once that is done, invalidate the cache entry on the current object and perform the actual table drop.  If this is wrapped in a transaction it will block any new table creates from before the external entries are invalidated, until after the drop is done.