Remove existence flag from SQLiteKeyStore
The optimization was causing some issues, and now the overall collection invalidation does not have to involve resetting key stores at all.  The collection itself can simply be invalidated and the next collection that wants to use the table will create it because the check is on the data file itself.