CBL-3419: Add collectionContext to C4DocumentEnded
:warning: API Change -> Added collectionContext to C4DocumentEnded

This change requires a lot of changes to other classes that interact with this struct.  The majority of the changes are to pass in the collection context where needed to ReplicatedRev (which is dynamically cast to C4DocumentEnded).  However, ChangesFeed needed to get a hold of the collection index to use, in order to know which context to add.