CBL-3406 : refactor for the separation of logics of active and passiv… (#1516)
CBL-3406 : refactor for the separation of logics of active and passive Replicators.
For the active replicators, workers, such as pushers and pullers, are created in Replicator's constructors. We know what to create from Replicators' configs. For the passive Replicators, we don't know what collections are going to participate until we receive the first message, getCollections. So, we will create all necessary workers after afterwards. Also, all messages come with collection index, and we will dispatch the message to the target worker based on the collection index.