CBL-3512: Fix some bugs regarding mismatched collections
For example, if the active side has more collections in the list than the passive one there needs to be a check that we are not running off the end of the passive side collection list.  Also clean up the code a bit to try to make it approachable for anyone seeing it for the first time.

Add in some mismatched collection test scenarios.