CBL-3009 : quoting collection names. (#1563)
By N1QL grammar, collection names and scope names are IDENTIFIERs, which can be quoted, just like the property names in a property path. That is, we may say, "SELECT ssn FROM `human.resource`.payroll", where "human.resouce" represents the scope name in the grammar, and "payroll" represents the collection name. Currently, however, the parser collapses them into one collection path, "human.resource.payroll". The original scope/collection names are not recoverable. Usually, this is not a problem, because no matter where you put the divide, the part with dot in it won't qualifiy as a scope or collection name in the later phase of the parser. However, the loss of the original intent of the N1QL expression can cause a problem. If unqualified -- not led by a scope name -- a collection name may represent the default collection. For example, if "human.resource.db" is the name of the database, "SELECT name FROM `human.resource.db` should mean to select the names from the default collection of the database.

Therefore,
1. we parse the collection path in N1QL expression into separate properties, "SCOPE" and "COLLECTION" in JSON.
2. We allow each properties, of SCOPE & COLLECTION, to be quoted, and save them in JSON string with escaped dot.
3. In the JSON form, we continue to allow omitting SCOPE and the value of COLLECTION to be dot-separated path, but being a separator, it must not be escaped.

If dot is not used in scope or collection names, there should not be any change from the current behavior. If they include dots, the parser will determine whether it corresponds to the database name and represent the default collection, or invalid collection names.