CBL-3698: Repeat the tests in ReplicatorCollectionTest.cc with Sync-g… (#1585)
* CBL-3698: Repeat the tests in ReplicatorCollectionTest.cc with Sync-gateway server.
Tests in ReplicatorCollectionTest.cc are performed with loopback connections between two replicators in the same executable. We port those tests with Sync-gateway server on the local machine, as appropriate to test the interaction between the replicator in CBL as client and Sync-gateway as the server.
The new tests are put in ReplicatorCollectionSGTest.cc. Still to be done is "Resolve Conflict", "Reset Checkpoint with Pull", and "Filters & docIDs with Multiple Collections".
* Cases remain to be done: "Resolve Conflicts", "Push & Pull". They are included in separate tickets.