CBL-3921: Error when sending _bulk_docs request with large number of docs
There are two occasions of errors:
1.  [TLS] ERROR: mbedTLS(C): mbedtls_ssl_send_alert_message() returned -80 (-0x0050)
This error is due to the fact that we send the request twice. First time we send it without proper auth-header, and SG may responds with 401/authentation without reading out the entire body of the request. At the socket level, it finds socket closed as it is finishing up the write. We can disregard this error in this case.
2. [WS] WARNING: ClientSocket got POSIX error 35 "Resource temporarily unavailable." This is actually due to socket timeout. I added warning in this case, so that we can decide whether to dial up the timeout or reduce the size of the bulk. To match the size (1000 docs) used in the corresponding test with Walrus, I needed to bump the timeout from default 5 seconds to 20 seconds.