CBL-4254: P2P tests giving invalid collection error (#1715)
The error occurs when the database name contains dots and URL parser takes the suffix after the dot as the collection, which is not valid in this context. Since the P2P protocol only uses the database, we simply take the whole string in the path as the name of the database which may contain dots.