CBL-4288: Data getting corrupted during PushPull collection replication (#1727)
* CBL-4288: Data getting corrupted during PushPull collection replication

This only happens on Windows, seemingly, and the only theory I have is that there is an optimization issue that causes some stale reads of a variable.  The details of the issue is that when reading the Fleece shared keys out of the "remote" db in a db to db pushpull replication, the number of keys in the copy of the shared keys object is 0 instead of what it should be.  This causes a faulty check for "unchanged" shared keys (memory address equal, and count > initial count) which in turn encodes incorrect keys onto the document.

This change seems to fix it, though it is unclear why.  The only theory I have is that reading the "count" variable first is forcing a re-evaluation which then uses the correct value when copying.

* Add in a retry loop, and give up if not successful

Once in several thousand runs, I saw the previous commit still fail.  Usually one loop of the retry is enough, but above all else if it fails for whatever reason then the writes need to stop.