CBL-4377: Not moving deleted docs when open database prior 3.1
With 3.1, as we open a database of old version, we will move all deleted docs to a separate database table. For huge database, say a database of gigabytes, it may take a long time. We are proposing here to skip the migration of the deleted docs in the default collections as the database is opened.
This change violated the post-condition after the database is opened that all the normal keystore does not contain deleted docs. We rollback the changes for the dual-keystore feature with regard to the default collections and adjust the affected test cases.

*warning* the violation of the above post condition broke the test, "Database Upgrade From 2.7 to Version Vectors", c.f. CBL 4382. I disabled it in this PR.