CBL-4377: Not moving deleted docs when open database prior 3.1 (#1741)
CBL-4377: Not moving deleted docs when open database prior 3.1

In Version 3.1, deleted documents are put in separate keystores. As we open a database of old version, we will move all deleted docs to the separate keystores. For huge database, e.g. a database of gigabytes, it may take unbearable minutes. We are proposing here to skip the migration of deleted docs for the default collections. (Notice that databases from previous versions do not have named collections.)
This change will violate the post-condition after the database is opened that the live keystore contains no deleted docs. Therefore, we rollback the changes for the dual-keystore feature with regard to the default collections and make necessary adjustments in the areas where those changes affect.

*warning* the violation of the above post condition broke the test, "Database Upgrade From 2.7 to Version Vectors", c.f. CBL 4382. I disabled it in this PR. The version vector has not been used.

Areas affected by the changes are: QueryParser and BothKeyStore. Test cases that are affected by QueryParser are adjusted. Targeted to changes in BothKeyStore, more scenarios are added to testOpeningOlderDBFixture(): including
* Check document counts
* Enumerate in descending order
* Enumerate only non-deleted doc
* Update and delete docs