CBL-4499: Replicator may get stuck when there is an error of "Invalid delta"
Replicator is stuck in busy state when there is an error thrown while applying delta to create full fleece doc

This PR does not fix the delta error. (The reason of the delta errors as shown in the ticket is still unknown.) Instead, I try to fix the problem of "stuck in busy state." In fact, our code already has paths to deal with the possible delta error. After injecting delta errors for push and pull, bugs emerged as the ticket describes.

After the fix, the delta error will trigger the second round of "rev" message which uses full rev as opposed to the delta rev.