CBL-4101: adapting tests based on SG 3.0 of walrus mode to also work with SG 3.1 backed by CB server.
When Lithium (aka v3.0) was released, all replicator tests were run against SG of the Walrus mode. With v3.1, we started to use SG backed up by actual CB servers. We ported tests in release/3.1, and the associated test framework, to lithium and put them in ReplicatorSGTest, and moved original tests based on Walrus to ReplicatorWalrusTest.
This commit adapts the original tests of Walrus to also work with SG(3.1) backed by CB server.