CBL-4533 : Add docker compose running CBS and SGs for tests (#1780)
The docker compose will start one CBS and two SGs : 3.1 SG with flowers collection configured (:4984/scratch) and legacy 3.0 SG with default collection configured (:4884/scratch-30). Found that when connecting two SGs with the same database name to the same CB Server, SG could get confused and result to collection or namespace not found error.

The SG's certs are copied from the top level cert directory to docker/sg/cert directory. This is mandatory as the docker cannot copy anything from the parent directory, and symlink is not supported either. So I made the top level cert directory a symlink of docker/sg/cert instead.

The default SG binaries set in docker-compose.yml are ARM64 based and the SSL is enabled. This could be adjusted. To reconfigure without changing docker-compose.yml, create .env with SG_DEB and SSL properties.

Tested by running all tests with "[.SyncServerCollection],[.SyncServer30]" tags. And with the docker compose, we could setup GH Action to run [.SyncServerCollection],[.SyncServer30] tests without much of the work (Not in this PR).

Cherry-picked @callumbirks 's 0706d97 commit from the docker-fix branch that updates the test code to use port 4884 for SyncServer30 tests.