Fix: Upgrade to VV setting incorrect version(s)
The code that assigns a version-vector to a doc during DB upgrade was
left unfinished in my earlier commits. I saw today it isn't creating
the right VV.

* The latest replicated revision now becomes a version whose timestamp
  is the minimum allowable value plus its old generation number.
  This is a valid timestamp, and provides an easy way to recover the
  generation number for e.g. display in `cblite`.
* If the current revision has not been replicated it becomes a new
  version tagged with the "me" source ID. But its timestamp needs to
  be valid: it now uses the current logical clock as usual.

While I was at it, since kMinValidTime is now more visible in upgraded
revIDs, I changed it to a rounder value, 0x1770000000000000. This is
in the recent past (sometime between June and now.)

We may refine this further during development, especially once SG
comes aboard, but this update at least doesn't create illegal versions
with too-small timestamps!