CBL-4382: Test "Database Upgrade From 2.7 to Version Vectors" Fails
Current, the signature KeyStore::set only permits insertion of new records, and hence must pick a new sequence (updateSequence). This is right for the top leve store, namely BothKeyStore. However, to move a record from sub-store to the other one, we delete it from one and intert it to the other one while keeping the sequence. The sub-store, or SQLiteKeyStore, also inherits from KeyStore. This is why we change the signature and make updateSequence and insert seaparate and ensure the integrity at appropriate places.