CBL-4898: Handling blobs that reference attachments that are already deleted in the remote.
If the rev pulled down from SG includes top level property "_attatchments", it should include all attachments that exist in the SG for the current revision. If a blob in the revision's body references an attachment that is not in "_attachments," it is bound for "getAttachment" to fail. We try to detect this case and, instead of handling the bound error response, err our within the Core library.