CBL-4898: Regression in pull of blobs/legacy attachment handling
This occurs when an attachment is deleted in SG but still used in a document's blob. SG gives error, HTTP/403, "Attachment's doc not being synced". We will not solve the root cause of the problem. Instead, we will log a clearer error in CBL, like: 'Blob in "att1" with digest "sha1-ERWD9RaGBqLSWOQ+96TZ6Kisjck=" is not in attachments.'.