Upgrade docs to version vectors lazily, on 1st write
Eliminates the cost of upgrading an existing db to version vectors.
Instead of converting all docs to version-vector format the first
time the db is opened, it leaves them alone.

A C4Document is opened using its existing versioning type, unless
the new document flag kDocGetUpgraded is used. (Even that flag
doesn't change the representation on disk.)

A C4Document is upgraded on disk when it's saved.

The replicator loads docs using kDocGetUpgraded since it uses
version vectors.