CBL-5030: Opening db of v2 into db of VersionVector causes conversion of all documents.
For the legacy database of huge size, we cannot afford to enumerate and convert all documents as we open it into the database of current versioning. Fortunately, we are able to handle documents of all earlier versions, specifically TreeVersioning_v2 and TreeVersioning, in the database of current VectorVersioning. The program will upgrade those documents to the current versioning at appropriate time on demand.