CBL-4547: Allow DictKeys to cache shared keys from query results (#1923)
Update to the enhancement in Fleece with regard to the SharedKeys.
For PersistentSharedKeys, the key (Dict::Key) can be cached only if its not in a trasaction. For the general SharedKeys, the default allows the key to be cashed (or shared). However, if it's possible that the underlying document may change after the key is used, client code can disable the caching. We want to be able to cache SharedKeys from the query results, but not for the tempSharedKeys in DBAccess.