CBL-5369: allKeyStoreNames() should ignore indexes
SQLiteKeyStore::allKeyStoreNames() had a very long-lasting bug wherein
it returned the names of tables that are used as indexes -- that
includes FTS, array indexes, predictive indexes, and now vector.

There was a band-aid fix for this (CBL-3824) in DatabaseImpl.cc
(keyStoreNameToCollectionSpec) which only filtered out FTS indexes;
others would get opened as KeyStores when the DB was reopened
 and cause errors.

I put in the proper fix (ignore tables with ":" in their names),
removed the band-aid, and added assertions in tests to make sure that
no type of index gets misinterpreted as a KeyStore.

Fixes CBL-5369