Test vector query with more-than-default number of results
Number of results was being capped at 8 by vectorsearch, due to a
virtual-table bug in SQLite. It wasn't detected because the tests
set the limit to 5. This commit increases it to 10, which detected
the problem and caused a test failure.

This has been fixed by updating SQLite to v3.43 in EE, in the
couchbase-lite-core-EE repo (commit 8f64a25a).