CBL-5444: Wrong vector dimensions shouldn't cause error saving doc
Fix is to pass the required dimensions to the SQLite function that
encodes the property, and have it return NULL if they don't match,
as it already does if the vector is missing or otherwise invalid.

Renamed that SQLite function from `fl_vector_value` to
`fl_vector_to_index`, to clarify that it's only used when indexing.