Added SQLite `carray` extension
This extension is part of the SQLite source tree but not built in.

Primary impetus for adding it is that the vector-search extension now
uses it, but as an extension can't bundle the code, so it requires
the owner of the SQLite handle to load it.

It's a great optimization for passing lots of values into a query in
an `IN(...)` clause, without having to encode every single value into
the SQL string. I added it to the SQLiteCpp library as well as C++
API for using it when binding parameters.

The one place we can immediately use it in LiteCore is
SQLiteKeyStore::withDocBodies(), so I updated that method. It might
give us a tiny boost in replication performance...