Disney CBSE 16662: FOTIFY: thread_mutex crash issue
Make C4QueryObserver ref-counted. We notify the observers outside mutex to avoid callbacks recurring back to the locked mutex. This makes observers liable being destroyed as whilst callback. We bump the ref-counts before leaving the mutex.

This change causes reference cycle when the observer is enabled. Client must disable (setEnabled(false)) to break the cycle and free it when done with it.