Part of CBL 5579 in order to facilitate VS on .NET Android
This probably affects Java Android as well, but at least on lower API levels a native library without the lib prefix will not be automatically extracted in legacy packaging.  So Android needs to look for libCouchbaseLiteVectorSearch while others continue looking for CouchbaseLiteVectorSearch.  In the meantime added more logging to show details about the loading process.