CBL-5346: Logging Replicator reasons of state change (#2016)
* CBL-5346: Logging Replicator reasons of state change

We follow the decision paths as new states are computed and log them by "reason=<reason>". The goal is, from the triple in the log, (objPath=, status=, reason=), we can unambiguously identify in the source code the place where the decision is reached, and the value of the variable that is responsible for the decision to transition to the state.