CBL-5646: Null dereference crash in gotHTTPResponse (#2020)
* CBL-5646: Null dereference crash in gotHTTPResponse

As BLIPIO receives HTTPResponse, it passes it upstream to Connection. When doing it, it calls gotHTTResponse on _connection without checking it against null. This is okay if it's called before it is closed, for we reset it to null when onClose is called. To guard against this situation, we check it before calling it.
To avoid racing with regard to _connection, we dispatch the method to the queue.