CBL-5481:  Lazy vector index updating (#2037)
* CBL-5481: Lazy vector index updating

* Lazy/async vector index updating [API]

- Adds the flag `lazy` to C4VectorIndexOptions. If true, the index
  will not update automatically, except when documents are deleted.
- The class C4LazyIndex represents a lazy vector index.
- The class C4LazyIndexUpdate is used to update the lazy index.
  It presents a list of Fleece values that are the result of the
  index's expression. The client must compute and store a vector for
  each one, then call its `finish` method to update the index.
- Also wrapped `#ifdef COUCHBASE_ENTERPRISE` around all VS API.

* Improved API, adding C4Index type

* CI fixes

* Added skipVectorAt method, and clarified API docs

* Some additions to SequenceSet, preparing for LazyIndex

- Added I/O methods for persisting to JSON or Fleece
- Moved most method bodies to new SequenceSet.cc
- SequenceSetTest.cc wasn't being built for some reason; fixed that.
- Moved a stray test out of ReplicatorAPITest.cc

---------

Co-authored-by: Jianmin Zhao <jianmin.zhao@couchbase.com>

* Support latest vectorsearch (dev branch) and hybrid queries.

* Added SQLite `carray` extension

This extension is part of the SQLite source tree but not built in.

Primary impetus for adding it is that the vector-search extension now
uses it, but as an extension can't bundle the code, so it requires
the owner of the SQLite handle to load it.

It's a great optimization for passing lots of values into a query in
an `IN(...)` clause, without having to encode every single value into
the SQL string. I added it to the SQLiteCpp library as well as C++
API for using it when binding parameters.

The one place we can immediately use it in LiteCore is
SQLiteKeyStore::withDocBodies(), so I updated that method. It might
give us a tiny boost in replication performance...

* Hybrid vector query support

- When vector_match() is the only criterion in the WHERE clause,
  OR if an explicit max_results arg is given, it's a "plain" query
  like already existed.
- Otherwise it's a "hybrid" query, which invokes the vectorsearch
  extension differently (with a JOIN constraint on its rowid column.)
  This is less efficient, but computes distances for all the rows
  selected by the other WHERE tests, instead of just finding the
  closest docs in the whole collection, so it gives more accurate
  results.

---------

Co-authored-by: Jianmin Zhao <jianmin.zhao@couchbase.com>

* Refer to submodule SQLiteCpp by branch 2.0.0-couchbase
Fix Xcode project files: carray.c to carray.cc, carray_bind.c to carray_bind.cc
Fix Cmake sources for the above files.
Disable GH check on BlackDuck.
!NO_BD_GITHUB

---------

Co-authored-by: Jens Alfke <jens@couchbase.com>