CBL-5784: Deadlock when setParentObjectRef is called (#2044)
The deadlock is caused by a logging of Error in setParentObjectRef. This error is logically a case of assertion failure. The problem is this function requires the mutex that WarnError also requires and and, therefore, got locked because of non-recursive mutex.

It turned out the error was caused by mismatch of library, but we don't want it be locked when the error occurs. I scoped the mutex and moved the WarnError out of the mutex scope.