CBL-5798: Passive replicator incorrectly errors on collection ID in the BLIP message (#2051)
* CBL-5798: Passive replicator incorrectly errors on collection ID in the BLIP message

For the active replicator, we set a flag to mark it as collection-unaware if the only collection that is in the configuration is the default collction. This is so that we can connect to 3.0 SG that is not aware of the collection.

Unnecessarily, we applied the above logic to the passive replicator as well, because we will determine collection awareness of the passive replicator from the first request it receives. It poses no problem if the active replicator is also derived from Replicator. For replicating client not derived from Replicator, this causes problem if the client has only one default collection and connet to the passive replicator derived from Replicator. In this case, the collection awareness does not match and violates internal assertion.

We fixed it by applying the above rule based on the collection config to the active replicator only.