Use same vector-index options struct as vectorsearch repo (#2058)
* Use same vector-index options struct as vectorsearch repo

Replaced IndexSpec::VectorOptions with vectorsearch::IndexSpec.

* Reimplemented SQLiteKeyStore::isIndexTrained()

using a query on the vector index that checks whether any vectors
have a bucket not equal to -1.

As part of this I had to add the `const` attribute to
`KeyStore::getIndex()`.

* c4Collection::getIndexesInfo includes vector options

c4Collection::getIndexesInfo now includes the vector index options,
under the key `vector_options`, encoded in a string in the same
format as the virtual-table args in the CREATE INDEX statement.

* C4Index API update (added getOptions!)

- Added c4Index::getOptions, which returns the complete vector or
  FTS options used to create the index.
- Also added c4Index methods getType, getQueryLanguage, getExpression,
  isTrained.
- Added the equivalent C functions.
- Moved index related functions from c4Collection.h to c4Index.h.
- In c4Index.h, moved the older database-based functions like
  c4db_createIndex to the bottom, bracketed by
  `#ifndef C4_STRICT_COLLECTION_API`.