CBL-5925: Regression from table migration during getIndex
Slightly complicated because there are several situations to deal with:

1. There are no indexes at all (new database).  This means that getIndex / getIndexes should basically return nothing and not attempt any table action.
2. There are indexes that were created before the lazy index introduction.  This means that getIndex / getIndexes need to modify the table before proceeding
3. There are indexes that were created after the lazy index introduction.  This means that getIndex / getIndexes should not attempt any table action either.

The previous attempt used "ensureIndexTableExists" which does enough for 2 here, but too much for 1 as it tries to take a transaction to create a new table and migrate potential previous data.  However, not modifying the table at all works for 1 but crashes 2 because it tries to use more columns than are returned.